Method modified: #realInverseSlotFor: anObject. ---------------------------------------------------------------------------------- The slots originally store the reverse slot object in a dictionary related with the type. Slots objects are created on command, and store circumstantial information such as the index in the instance variables array. This information is meant to get out-dated when what-ever structural changes happen (such as modifying trait composition order, adding an instance variable in a class or superclass, etc).
For avoiding this problem of out-dating we changed the implementation to store in the cache a message send object. 

This could eventually penalize the execution time  performance. Never the less the other oportunity is to flush a cache on all the instances of a many classes. This could be quite complex in the case of a complex model with many instances (a common scenario in the context of moose)