Fix class generation
Fame has a code generator based on the refactoring model. But the way to generate classes evolved in Pharo 12 to go from class definitions as string to the use of the ShiftClassBuilder. This adds a compatibility to P12 about that and also rename some things to go from the #category name to #package